Electrum: move wallet db into the electrum package
